Initial tasks, with memory regions per task

+++ kernel/arch/amd64/memory/vmm.h
@@ -1,7 +1,8 @@
 #pragma once
 
 #include <arch/common/memory/vmm.h>
+#include <stddef.h>
 #include <stdint.h>
 
 void vmm_load(uintptr_t address);
-uint64_t vmm_get_direct_map_base();
+uintptr_t vmm_get_direct_map_base();
